services/websocket-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
redis==5.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aioredis==2.0.1
